requirements.txt
deriv-api
google-generativeai
asyncio
